DVX-6600 Extract query for common use
To be able to use it from both `Checker` and `Collector` we need a
common interface. This interface also allows to improve the tests for
query function without need to alter the contents of the objects.
