+++ pkg/dnsbl/common_test.go
@@ -39,3 +39,43 @@ func TestReverseAddress(t *testing.T) {
 		)
 	}
 }
+
+type mockLookup struct{}
+
+func (m mockLookup) lookup(address string) ([]string, error) {
+	var addresses []string
+	if address == "1.0.0.127.positive.dnsbl.com" {
+		addresses = []string{"127.0.0.1"}
+	} else {
+		addresses = []string{}
+	}
+	return addresses, nil
+}
+
+func TestQuery(t *testing.T) {
+	var table = []struct {
+		address string
+		result  int
+	}{
+		{
+			address: "1.0.0.127.positive.dnsbl.com",
+			result:  1,
+		},
+		{
+			address: "1.0.0.127.negative.dnsbl.com",
+			result:  0,
+		},
+	}
+	checker := mockLookup{}
+	for _, tc := range table {
+		result := query(checker, tc.address)
+		if result != tc.result {
+			t.Errorf(
+				"Result for %s should be %d instead of %d",
+				tc.address,
+				tc.result,
+				result,
+			)
+		}
+	}
+}
